Return gitter stream API
The first implementation of communication with Gitter was done by using
Streaming API. Later, it was replaced by reqular Rest API due to
problems with timeouts. Unfortunately, usage of Rest API was a bad idea,
since it makes too manu requests and Gitter server blocks the client.

This patch returns the communication by streaming api with a fix for
timeout issue: the default timeout of aiohttp session is quite small and
redundant, since gitter api sends empty messages to keep the
communication alive.